hyc_6-30_00:56

+++ src/main/java/com/jxdinfo/hussar/inisalary/controller/TInisalaryLevelController.java
@@ -1,13 +1,17 @@
 package com.jxdinfo.hussar.inisalary.controller;
 
+import com.alibaba.fastjson.JSONArray;
 import com.baomidou.mybatisplus.mapper.EntityWrapper;
 import com.baomidou.mybatisplus.mapper.Wrapper;
 import com.baomidou.mybatisplus.plugins.Page;
 import com.jxdinfo.hussar.core.base.controller.BaseController;
 import com.jxdinfo.hussar.inisalary.model.TInisalaryInfo;
+import com.jxdinfo.hussar.inisalary.model.TInisalaryInforLog;
+import com.jxdinfo.hussar.inisalary.service.ITDepartmentService;
 import com.jxdinfo.hussar.inisalary.service.ITInisalaryInfoService;
-import com.mysql.cj.x.protobuf.Mysqlx;
-import org.apache.ibatis.jdbc.Null;
+
+import com.jxdinfo.hussar.inisalary.service.ITInisalaryInforLogService;
+import com.jxdinfo.salary.PermissionManagement.controller.UtilController;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseBody;
@@ -22,12 +26,8 @@ import com.jxdinfo.hussar.core.log.type.BussinessLogType;
 import com.jxdinfo.hussar.inisalary.model.TInisalaryLevel;
 import com.jxdinfo.hussar.inisalary.service.ITInisalaryLevelService;
 
-import javax.servlet.http.HttpServletResponse;
-import java.io.IOException;
-import java.util.Arrays;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
+import java.sql.Timestamp;
+import java.util.*;
 
 /**
  * 初始工资维护控制器
@@ -46,6 +46,15 @@ public class TInisalaryLevelController extends BaseController {
 
     @Autowired
     private ITInisalaryInfoService  tInisalaryInfoService;
+
+    @Autowired
+    private ITInisalaryInforLogService tInisalaryInforLogService;
+
+    @Autowired
+    private UtilController utilController;
+
+    @Autowired
+    private ITDepartmentService departmentService;
 
     /**
      * 跳转到初始工资维护首页
@@ -112,15 +121,32 @@ public class TInisalaryLevelController extends BaseController {
     @RequiresPermissions("tInisalaryLevel:add")
     @ResponseBody
     public Object add(TInisalaryLevel tInisalaryLevel) {
-        TInisalaryLevel t = tInisalaryLevelService.selectByTwo(tInisalaryLevel.getDepartmentId()+"",tInisalaryLevel.getLevel()+"");
-        if(tInisalaryLevel.getMaximumAge()<tInisalaryLevel.getMinimumAge()) return -1;
-        if(t!= null) {
-            return 0;
+        //获取权限
+        String depName = departmentService.selectById(tInisalaryLevel.getDepartmentId()).getDepartmentName();
+        System.out.println("======"+utilController.getPermission("维护工资管理",depName)+"==========");
+        if(utilController.getPermission("维护工资管理",depName).equals(true)){
+            TInisalaryLevel t = tInisalaryLevelService.selectByTwo(tInisalaryLevel.getDepartmentId()+"",tInisalaryLevel.getLevel()+"");
+            if(tInisalaryLevel.getMaximumAge()<tInisalaryLevel.getMinimumAge()) return -1;
+            if(t!= null) {
+                Map<String,Object> res= new HashMap<>();
+                res.put("code",500);
+                res.put("message","该部门等级已存在表中");
+                return res;
+            }
+            else{
+                tInisalaryLevelService.insert(tInisalaryLevel);
+                Map<String,Object> res= new HashMap<>();
+                res.put("code",200);
+                res.put("message","添加成功");
+                return res;
+            }
+        }else{
+            Map<String,Object> res= new HashMap<>();
+            res.put("code",500);
+            res.put("message","没有权限操作");
+            return res;
         }
-        else{
-            tInisalaryLevelService.insert(tInisalaryLevel);
-            return 1;
-        }
+
 
     }
 
@@ -131,14 +157,57 @@ public class TInisalaryLevelController extends BaseController {
     @BussinessLog(key = "/tInisalaryLevel/delete", type = BussinessLogType.DELETE, value = "删除初始工资维护")
     @RequiresPermissions("tInisalaryLevel:delete")
     @ResponseBody
-    public Object delete(@RequestParam String tInisalaryLevelId,@RequestParam String tInisalaryLevelId2) {
+    public Object delete(@RequestParam List<String> tInisalaryLevelId,@RequestParam List<String> tInisalaryLevelId2) {
         //List idtInisalaryLevel = Arrays.asList(tInisalaryLevelId.split(","));
-        Map<String, Object> columnMap = new HashMap<String, Object>();
-        columnMap.put("DEPARTMENT_ID",tInisalaryLevelId);
-        columnMap.put("LEVEL",tInisalaryLevelId2);
 
-        tInisalaryLevelService.deleteByMap(columnMap);
-        return SUCCESS_TIP;
+//        System.out.println("==========="+s+"==============");
+//        System.out.println("==========="+tInisalaryLevelId.get(0)+"==============");
+//        System.out.println("==========="+tInisalaryLevelId2.get(0)+"==============");
+        int len = tInisalaryLevelId.size();
+        System.out.println("==========="+len+"==============");
+        boolean hasPermission = true;
+        for(int i=0;i<len;i++){ //先遍历判断是否有权限
+           // Map<String, Object> columnMap = new HashMap<String, Object>();
+            //获取权限
+            String id = tInisalaryLevelId.get(i).replaceAll("[^\\d]", "");
+           // String lev = tInisalaryLevelId2.get(i).replaceAll("[^\\d]", "");
+            String depName = departmentService.selectById(Integer.parseInt(id)).getDepartmentName();
+            System.out.println("======"+utilController.getPermission("维护工资管理",depName)+"==========");
+
+            if(utilController.getPermission("维护工资管理",depName).equals(true)){
+//                columnMap.put("DEPARTMENT_ID",id);
+//                columnMap.put("LEVEL",lev);
+//                tInisalaryLevelService.deleteByMap(columnMap);
+            }else{
+                hasPermission =false;//没权限
+            }
+
+        }
+        if(hasPermission){//有权限则删除
+            for(int i=0;i<len;i++){
+                Map<String, Object> columnMap = new HashMap<String, Object>();
+                //获取权限
+                String id = tInisalaryLevelId.get(i).replaceAll("[^\\d]", "");
+                String lev = tInisalaryLevelId2.get(i).replaceAll("[^\\d]", "");
+
+                columnMap.put("DEPARTMENT_ID",id);
+                columnMap.put("LEVEL",lev);
+                tInisalaryLevelService.deleteByMap(columnMap);
+
+            }
+            Map<String,Object> res= new HashMap<>();
+            res.put("code",200);
+            res.put("message","删除成功");
+            return res;
+        }else{
+            System.out.println("======没有权限aa========");
+            Map<String,Object> res= new HashMap<>();
+            res.put("code",500);
+            res.put("message","没有权限操作");
+            return res;
+        }
+
+
     }
 
 
@@ -150,26 +219,57 @@ public class TInisalaryLevelController extends BaseController {
     @RequiresPermissions("tInisalaryLevel:update")
     @ResponseBody
     public Object update(TInisalaryLevel tInisalaryLevel) {
-        String t1 =tInisalaryLevel.getDepartmentId()+"";
-        String t2 =tInisalaryLevel.getLevel()+"";
-        TInisalaryLevel tInisalaryLevel1 = tInisalaryLevelService.selectByTwo(t1,t2);
-        LogObjectHolder.me().set(tInisalaryLevel1);
-        Wrapper<TInisalaryLevel> wrapper = new EntityWrapper<>();
-        wrapper.eq("DEPARTMENT_ID",tInisalaryLevel.getDepartmentId());
-        wrapper.eq("LEVEL",tInisalaryLevel.getLevel());
-        tInisalaryLevelService.update(tInisalaryLevel,wrapper);
+        //获取权限
+        String depName = departmentService.selectById(tInisalaryLevel.getDepartmentId()).getDepartmentName();
+        System.out.println("======"+utilController.getPermission("维护工资管理",depName)+"==========");
 
-        //修改员工初始工资表
-        Map<String, Object> columnMap = new HashMap<String, Object>();
-        columnMap.put("DEPARTMENT_ID",tInisalaryLevel.getDepartmentId());
-        columnMap.put("LEVEL",tInisalaryLevel.getLevel());
-        List<TInisalaryInfo> infoList = tInisalaryInfoService.selectByMap(columnMap);
-        for (TInisalaryInfo t : infoList) {
-            t.setInitialSalary(tInisalaryLevel.getInitialSalary());
-            tInisalaryInfoService.updateById(t);
+        if(utilController.getPermission("维护工资管理",depName).equals(true)){
+            String t1 =tInisalaryLevel.getDepartmentId()+"";
+            String t2 =tInisalaryLevel.getLevel()+"";
+            int inisal1 =tInisalaryLevel.getInitialSalary();
+            TInisalaryLevel tInisalaryLevel1 = tInisalaryLevelService.selectByTwo(t1,t2);
+            int inisal2 = tInisalaryLevel1.getInitialSalary();
+            boolean equal =(inisal1==inisal2);
+            //    System.out.println(inisal1+"--------------"+inisal2+"-----------"+equal);
+
+            LogObjectHolder.me().set(tInisalaryLevel1);
+            Wrapper<TInisalaryLevel> wrapper = new EntityWrapper<>();
+            wrapper.eq("DEPARTMENT_ID",tInisalaryLevel.getDepartmentId());
+            wrapper.eq("LEVEL",tInisalaryLevel.getLevel());
+            tInisalaryLevelService.update(tInisalaryLevel,wrapper);
+
+            //修改员工初始工资表
+            Map<String, Object> columnMap = new HashMap<String, Object>();
+            columnMap.put("DEPARTMENT_ID",tInisalaryLevel.getDepartmentId());
+            columnMap.put("LEVEL",tInisalaryLevel.getLevel());
+            List<TInisalaryInfo> infoList = tInisalaryInfoService.selectByMap(columnMap);
+
+            for (TInisalaryInfo t : infoList) {
+                Date date = new Date();
+                Timestamp nousedate = new Timestamp(date.getTime());
+                TInisalaryInforLog tInisalaryInforLog = new TInisalaryInforLog(t.getStaffId(),t.getStaffName(),t.getDepartmentId(),t.getLevel(),
+                        t.getInitialSalary(),nousedate,0);
+                if(!equal){
+                    t.setInitialSalary(tInisalaryLevel.getInitialSalary());
+                    tInisalaryInfoService.updateById(t);
+                    tInisalaryInforLogService.insert(tInisalaryInforLog);
+                }else{
+                    t.setInitialSalary(tInisalaryLevel.getInitialSalary());
+                    tInisalaryInfoService.updateById(t);
+                }
+
+            }
+            Map<String,Object> res= new HashMap<>();
+            res.put("code",200);
+            res.put("message","修改成功");
+            return res;
+        }else{
+            Map<String,Object> res= new HashMap<>();
+            res.put("code",500);
+            res.put("message","没有权限操作");
+            return res;
         }
 
-        return SUCCESS_TIP;
     }
 
     /**
